SWM-103 fix: 가입 여부 확인 api 이동(member -> auth)
