+++ app/src/main/java/com/betterlife/antifragile/data/remote/AuthApiService.kt
@@ -7,12 +7,16 @@ import com.betterlife.antifragile.data.model.auth.response.AuthLoginResponse
 import com.betterlife.antifragile.data.model.auth.response.AuthReIssueTokenResponse
 import com.betterlife.antifragile.data.model.auth.response.AuthSignUpResponse
 import com.betterlife.antifragile.data.model.base.BaseResponse
+import com.betterlife.antifragile.data.model.enums.LoginType
+import com.betterlife.antifragile.data.model.member.response.MemberExistenceResponse
 import okhttp3.MultipartBody
 import retrofit2.Response
 import retrofit2.http.Body
+import retrofit2.http.GET
 import retrofit2.http.Multipart
 import retrofit2.http.POST
 import retrofit2.http.Part
+import retrofit2.http.Query
 
 /**
  * 인증 API 엔드포인트를 정의하는 인터페이스
@@ -35,4 +39,10 @@ interface AuthApiService {
     suspend fun reIssueToken(
         @Body request: AuthReIssueTokenRequest
     ): Response<BaseResponse<AuthReIssueTokenResponse>>
+
+    @GET("/members/existence")
+    suspend fun checkMemberExistence(
+        @Query("email") email: String,
+        @Query("loginType") loginType: LoginType
+    ): Response<BaseResponse<MemberExistenceResponse>>
 }
