+++ app/src/main/java/com/betterlife/antifragile/data/remote/MemberApiService.kt
@@ -1,11 +1,9 @@
 package com.betterlife.antifragile.data.remote
 
 import com.betterlife.antifragile.data.model.base.BaseResponse
-import com.betterlife.antifragile.data.model.enums.LoginType
 import com.betterlife.antifragile.data.model.member.request.MemberProfileModifyRequest
 import com.betterlife.antifragile.data.model.member.response.MemberDetailResponse
 import com.betterlife.antifragile.data.model.member.response.MemberMyPageResponse
-import com.betterlife.antifragile.data.model.member.response.MemberExistenceResponse
 import com.betterlife.antifragile.data.model.member.response.MemberProfileModifyResponse
 import com.betterlife.antifragile.data.model.member.response.MemberRemainNumberResponse
 import com.betterlife.antifragile.data.model.member.response.NicknameDuplicateResponse
@@ -36,12 +34,6 @@ interface MemberApiService {
 
     @GET("/members/remain-recommend-number")
     suspend fun getRemainRecommendNumber(): Response<BaseResponse<MemberRemainNumberResponse>>
-
-    @GET("/members/existence")
-    suspend fun checkMemberExistence(
-        @Query("email") email: String,
-        @Query("loginType") loginType: LoginType
-    ): Response<BaseResponse<MemberExistenceResponse>>
 
     @GET("/members/duplication-check")
     suspend fun checkNicknameExistence(
