+++ app/src/main/java/com/betterlife/antifragile/data/repository/AuthRepository.kt
@@ -7,6 +7,8 @@ import com.betterlife.antifragile.data.model.auth.response.AuthLoginResponse
 import com.betterlife.antifragile.data.model.auth.response.AuthReIssueTokenResponse
 import com.betterlife.antifragile.data.model.auth.response.AuthSignUpResponse
 import com.betterlife.antifragile.data.model.base.BaseResponse
+import com.betterlife.antifragile.data.model.enums.LoginType
+import com.betterlife.antifragile.data.model.member.response.MemberExistenceResponse
 import com.betterlife.antifragile.data.remote.AuthApiService
 import okhttp3.MultipartBody
 
@@ -25,6 +27,14 @@ class AuthRepository(
     suspend fun login(request: AuthLoginRequest): BaseResponse<AuthLoginResponse> {
         return safeApiCall {
             authApiService.login(request)
+        }
+    }
+
+    suspend fun checkMemberExistence(
+        email: String, loginType: LoginType
+    ): BaseResponse<MemberExistenceResponse> {
+        return safeApiCall {
+            authApiService.checkMemberExistence(email, loginType)
         }
     }
