+++ app/src/main/java/com/betterlife/antifragile/data/repository/MemberRepository.kt
@@ -1,11 +1,9 @@
 package com.betterlife.antifragile.data.repository
 
 import com.betterlife.antifragile.data.model.base.BaseResponse
-import com.betterlife.antifragile.data.model.enums.LoginType
 import com.betterlife.antifragile.data.model.member.request.MemberProfileModifyRequest
 import com.betterlife.antifragile.data.model.member.response.MemberDetailResponse
 import com.betterlife.antifragile.data.model.member.response.MemberMyPageResponse
-import com.betterlife.antifragile.data.model.member.response.MemberExistenceResponse
 import com.betterlife.antifragile.data.model.member.response.MemberProfileModifyResponse
 import com.betterlife.antifragile.data.model.member.response.MemberRemainNumberResponse
 import com.betterlife.antifragile.data.model.member.response.NicknameDuplicateResponse
@@ -40,14 +38,6 @@ class MemberRepository(
     suspend fun getRemainRecommendNumber(): BaseResponse<MemberRemainNumberResponse> {
         return safeApiCall {
             memberApiService.getRemainRecommendNumber()
-        }
-    }
-
-    suspend fun checkMemberExistence(
-        email: String, loginType: LoginType
-    ): BaseResponse<MemberExistenceResponse> {
-        return safeApiCall {
-            memberApiService.checkMemberExistence(email, loginType)
         }
     }
