+++ app/src/main/java/com/betterlife/antifragile/presentation/ui/auth/viewmodel/LoginViewModelFactory.kt
@@ -4,21 +4,16 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.betterlife.antifragile.config.RetrofitInterface
 import com.betterlife.antifragile.data.repository.AuthRepository
-import com.betterlife.antifragile.data.repository.MemberRepository
 
-class LoginViewModelFactory(
-    private val token: String,
-) : ViewModelProvider.Factory {
+class LoginViewModelFactory : ViewModelProvider.Factory {
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
             val authApiService = RetrofitInterface.createAuthApiService()
             val authRepository = AuthRepository(authApiService)
-            val memberApiService = RetrofitInterface.createMemberApiService(token)
-            val memberRepository = MemberRepository(memberApiService)
 
             @Suppress("UNCHECKED_CAST")
-            return LoginViewModel(authRepository, memberRepository) as T
+            return LoginViewModel(authRepository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
